Test.Cardano.Ledger.Alonzo.AlonzoEraGen

Description

AlonzoEra instances for EraGen and ScriptClass

Documentation

vKeyLockedAdaOnly ∷ Crypto c ⇒ TxOut (AlonzoEra c) → Bool

We are choosing new TxOut to pay fees, We want only Key locked addresss with Ada only values.

phase2scripts3Arg ∷ ∀ era. AlonzoEraScript era ⇒ [TwoPhase3ArgInfo era]

phase2scripts2Arg ∷ ∀ era. AlonzoEraScript era ⇒ [TwoPhase2ArgInfo era]

phase2scripts3ArgSucceeds ∷ ∀ era. AlonzoEraScript era ⇒ Script era → Bool

phase2scripts2ArgSucceeds ∷ ∀ era. AlonzoEraScript era ⇒ Script era → Bool

genPlutus2Arg ∷ AlonzoEraScript era ⇒ Gen (Maybe (TwoPhase2ArgInfo era))

genAlonzoMint ∷ Crypto c ⇒ MultiAsset c → Gen (MultiAsset c, [AlonzoScript (AlonzoEra c)])

Gen a Mint value in the Alonzo Era, with a 10% chance that it includes an AlonzoScript

genPair ∷ Gen a → Gen b → Gen (a, b)

genPlutusData ∷ Gen Data

genSet ∷ Ord a ⇒ Gen a → Gen (Set a)

genAux ∷ ∀ c. Mock c ⇒ Constants → Gen (StrictMaybe (AlonzoTxAuxData (AlonzoEra c)))

unTime ∷ AlonzoScript era → Timelock era

okAsCollateral ∷ ∀ c. Mock c ⇒ UTxO (AlonzoEra c) → TxIn c → Bool

genAlonzoTxBody ∷ ∀ c. Mock c ⇒ GenEnv (AlonzoEra c) → UTxO (AlonzoEra c) → PParams (AlonzoEra c) → SlotNo → Set (TxIn c) → StrictSeq (TxOut (AlonzoEra c)) → StrictSeq (TxCert (AlonzoEra c)) → Withdrawals c → Coin → StrictMaybe (Update (AlonzoEra c)) → StrictMaybe (AuxiliaryDataHash c) → Gen (TxBody (AlonzoEra c), [Script (AlonzoEra c)])

genSlotAfter ∷ SlotNo → Gen SlotNo

genAlonzoPParamsUpdate ∷ ∀ c. Crypto c ⇒ Constants → PParams (AlonzoEra c) → Gen (PParamsUpdate (AlonzoEra c))

Gen an Alonzo PParamsUpdate, by adding to a Shelley PParamsData

genAlonzoPParams ∷ ∀ c. Crypto c ⇒ Constants → Gen (PParams (AlonzoEra c))

bigMem ∷ Natural

bigStep ∷ Natural

genMaxTxExUnits ∷ Gen ExUnits

genMaxBlockExUnits ∷ Gen ExUnits

sumCollateral ∷ (EraTx era, AlonzoEraTxBody era) ⇒ Tx era → UTxO era → Coin

storageCost ∷ ∀ era t. (EraPParams era, EncCBOR t) ⇒ Integer → PParams era → t → Coin

addRedeemMap ∷ ∀ c. Crypto c ⇒ (Data, Natural, Natural) → AlonzoPlutusPurpose AsIxItem (AlonzoEra c) → Map (AlonzoPlutusPurpose AsIx (AlonzoEra c)) (Data (AlonzoEra c), ExUnits) → Map (AlonzoPlutusPurpose AsIx (AlonzoEra c)) (Data (AlonzoEra c), ExUnits)

getDataMap ∷ ∀ era. Era era ⇒ ScriptInfo era → Map (ScriptHash (EraCrypto era)) (Script era) → Map (DataHash (EraCrypto era)) (Data era)

dataFromAddr ∷ ∀ c. Mock c ⇒ Addr c → StrictMaybe (DataHash c)

If an Address is script address, we can find a potential data hash for it from genEraTwoPhase3Arg, which contains all known 3 arg plutus scripts in the tests set. If the script has is not in that map, then its data hash is SNothing.

dataMapFromTxOut ∷ ∀ c. Mock c ⇒ [TxOut (AlonzoEra c)] → TxDats (AlonzoEra c) → TxDats (AlonzoEra c)

We can find the data associated with the data hashes in the TxOuts, since genEraTwoPhase3Arg, which contains all known 3 arg plutus scripts stores the data.

addMaybeDataHashToTxOut ∷ Mock c ⇒ TxOut (AlonzoEra c) → TxOut (AlonzoEra c)

someLeaf ∷ ∀ era. (AllegraEraScript era, NativeScript era ~ Timelock era) ⇒ Proxy era → KeyHash 'Witness (EraCrypto era) → AlonzoScript era

langsUsed ∷ AlonzoEraScript era ⇒ Map (ScriptHash (EraCrypto era)) (Script era) → Set Language

given the "txscripts" field of the TxWits, compute the set of languages used in a transaction

Orphan instances

Mock c ⇒ EraGen (AlonzoEra c)

Instance details

Methods

genGenesisValue ∷ GenEnv (AlonzoEra c) → Gen (Value (AlonzoEra c))

genEraTwoPhase3Arg ∷ [TwoPhase3ArgInfo (AlonzoEra c)]

genEraTwoPhase2Arg ∷ [TwoPhase2ArgInfo (AlonzoEra c)]

genEraTxBody ∷ GenEnv (AlonzoEra c) → UTxO (AlonzoEra c) → PParams (AlonzoEra c) → SlotNo → Set (TxIn (EraCrypto (AlonzoEra c))) → StrictSeq (TxOut (AlonzoEra c)) → StrictSeq (TxCert (AlonzoEra c)) → Withdrawals (EraCrypto (AlonzoEra c)) → Coin → StrictMaybe (Update (AlonzoEra c)) → StrictMaybe (AuxiliaryDataHash (EraCrypto (AlonzoEra c))) → Gen (TxBody (AlonzoEra c), [Script (AlonzoEra c)])

genEraAuxiliaryData ∷ Constants → Gen (StrictMaybe (TxAuxData (AlonzoEra c)))

updateEraTxBody ∷ UTxO (AlonzoEra c) → PParams (AlonzoEra c) → TxWits (AlonzoEra c) → TxBody (AlonzoEra c) → Coin → Set (TxIn (EraCrypto (AlonzoEra c))) → TxOut (AlonzoEra c) → TxBody (AlonzoEra c)

addInputs ∷ TxBody (AlonzoEra c) → Set (TxIn (EraCrypto (AlonzoEra c))) → TxBody (AlonzoEra c)

genEraPParamsUpdate ∷ Constants → PParams (AlonzoEra c) → Gen (PParamsUpdate (AlonzoEra c))

genEraPParams ∷ Constants → Gen (PParams (AlonzoEra c))

genEraTxWits ∷ (UTxO (AlonzoEra c), TxBody (AlonzoEra c), ScriptInfo (AlonzoEra c)) → Set (WitVKey 'Witness (EraCrypto (AlonzoEra c))) → Map (ScriptHash (EraCrypto (AlonzoEra c))) (Script (AlonzoEra c)) → TxWits (AlonzoEra c)

genEraGoodTxOut ∷ TxOut (AlonzoEra c) → Bool

constructTx ∷ TxBody (AlonzoEra c) → TxWits (AlonzoEra c) → StrictMaybe (TxAuxData (AlonzoEra c)) → Tx (AlonzoEra c)

genEraScriptCost ∷ PParams (AlonzoEra c) → Script (AlonzoEra c) → Coin

genEraDone ∷ UTxO (AlonzoEra c) → PParams (AlonzoEra c) → Tx (AlonzoEra c) → Gen (Tx (AlonzoEra c))

genEraTweakBlock ∷ PParams (AlonzoEra c) → Seq (Tx (AlonzoEra c)) → Gen (Seq (Tx (AlonzoEra c)))

hasFailedScripts ∷ Tx (AlonzoEra c) → Bool

feeOrCollateral ∷ Tx (AlonzoEra c) → UTxO (AlonzoEra c) → Coin

Mock c ⇒ MinGenTxout (AlonzoEra c)

Instance details

Methods

calcEraMinUTxO ∷ TxOut (AlonzoEra c) → PParams (AlonzoEra c) → Coin

addValToTxOut ∷ Value (AlonzoEra c) → TxOut (AlonzoEra c) → TxOut (AlonzoEra c)

genEraTxOut ∷ GenEnv (AlonzoEra c) → Gen (Value (AlonzoEra c)) → [Addr (EraCrypto (AlonzoEra c))] → Gen [TxOut (AlonzoEra c)]

Crypto c ⇒ ScriptClass (AlonzoEra c)

Instance details

Methods

basescript ∷ Proxy (AlonzoEra c) → KeyHash 'Witness (EraCrypto (AlonzoEra c)) → Script (AlonzoEra c)

isKey ∷ Proxy (AlonzoEra c) → Script (AlonzoEra c) → Maybe (KeyHash 'Witness (EraCrypto (AlonzoEra c)))

isOnePhase ∷ Proxy (AlonzoEra c) → Script (AlonzoEra c) → Bool

quantify ∷ Proxy (AlonzoEra c) → Script (AlonzoEra c) → Quantifier (Script (AlonzoEra c))

unQuantify ∷ Proxy (AlonzoEra c) → Quantifier (Script (AlonzoEra c)) → Script (AlonzoEra c)